FeedbackValueItem clean-up
